perf(dashboard): lazy-load recharts on metrics + sparkloop pages (#69)
Extract recharts JSX from 3 dashboard client components into sibling
wrapper files, and replace inline charts with `next/dynamic` imports.
Recharts (~400 KB minified) now ships in a separate chunk loaded only
when the chart actually mounts, instead of in the initial page JS.

Wrapper files (created):
- metrics/MetricChart.tsx
- sparkloop/components/OffersChart.tsx
- sparkloop/components/OverviewChart.tsx

Loading placeholders use `role="status"` + `aria-busy="true"` for
screen readers (WCAG 4.1.3). Heights match the rendered chart's
ResponsiveContainer height pixel-for-pixel to prevent CLS.

Build output:
- /dashboard/[slug]/metrics: 2.8 kB / 118 kB First Load JS
- /dashboard/[slug]/sparkloop: 19.2 kB / 134 kB First Load JS

Scope deliberately recharts-only. The other two libs called out in
issue #69 (`@dnd-kit`, `react-image-crop`) need dedicated PRs:
- @dnd-kit is hooks-driven across 22 files / 55 imports — extracting
  drag-and-drop trees behind dynamic boundaries is a real refactor
- react-image-crop has 18 importers using `<ReactCrop>` — doable but
  separate scope
